# sdks/sdkwork-im-sdk/openapi/craw-chat-im.sdkgen.yaml
openapi: 3.1.0
info:
  title: Sdkwork IM IM Standardized Development API
  version: 0.1.0
  description: >-
    IM standardized development OpenAPI contract for conversations, messages, realtime, media, streams, and social IM
    flows.
...
paths:
...
  /im/v3/api/chat/conversations/{conversationId}/agent_handoff/resolve:
    parameters:
      - $ref: '#/components/parameters/ConversationIdPath'
    post:
      tags:
        - chat
      operationId: conversations.agentHandoff.resolve
      summary: Resolve agent handoff
      parameters:
        - $ref: '#/components/parameters/ConversationIdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '400':
          description: HTTP 400 problem
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: HTTP 401 problem
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: HTTP 403 problem
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: HTTP 404 problem
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
      security:
        - AuthToken: []
          AccessToken: []
      x-sdkwork-owner: sdkwork-im
      x-sdkwork-api-authority: sdkwork-im.im
...
components:
  parameters:
    ConversationIdPath:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    AckResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
      required:
        - ok
...
    ProblemDetail:
      type: object
      additionalProperties: true
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
        retryable:
          type: boolean
...
  securitySchemes:
    AuthToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AccessToken:
      type: apiKey
      in: header
      name: Access-Token
...
security:
  - AuthToken: []
    AccessToken: []
...